// src/app/model/MvtStock.ts
export class MvtStock {
  dateMvt: Date;
  idCaisseEn: string;
  idFournisseur: string;
  idMvt: string;
  idPointVenteDest: string;
  idPointVenteSource: string;
  idProduitProintVente: string;
  idUser: string;
  motif: string;
  quantite: number;
  sens: string;
  type: string;
  valeur: number;
  isConversion: boolean;
  mesureConversion: string;
  qteConversion: number;
  prixVente: number;
  isGenereBail: boolean;
  tva: number;
}